Reoganize procedureSyntax tests
+ adapt scalacOptions for scala 3

Update docs/rules/ProcedureSyntax.md
