+++ scalafix-tests/input/src/main/scala-2/test/procedureSyntax/ProcedureSyntax.scala
@@ -1,7 +1,7 @@
 /*
 rules = ProcedureSyntax
  */
-package test
+package test.procedureSyntax
 
 object ProcedureSyntax {
   // This is a comment
