+++ scalafix-tests/output/src/main/scala-2/test/procedureSyntax/ConstructorProcedureSyntax.scala
@@ -1,4 +1,4 @@
-package test
+package test.procedureSyntax
 
 class ConstructorProcedureSyntax(i: Long) {
   def this(a: Int) = { this(0L + 1) }
