+++ scalafix-tests/output/src/main/scala-2/test/procedureSyntax/ProcedureSyntax.scala
@@ -1,4 +1,4 @@
-package test
+package test.procedureSyntax
 
 object ProcedureSyntax {
   // This is a comment
